requirements.txt
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
statsmodels==0.12.1
matplotlib==3.3.3

#sklearn==0.23
scikit-learn>=0.22.0,<0.23
auto-sklearn==0.11.1

numpy==1.18.5
pandas==1.1.4
Cython==0.29.21

selenium==3.141.0
parsel==1.6.0


tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tornado==6.1
tqdm==4.50.2
threadpoolctl==2.1.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4

tqdm==4.50.2
gensim==3.8.3
#fasttext==2.6.0

threadpoolctl==2.1.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
testresources==2.0.1
texttable==1.6.2

bcrypt==3.1.7
beautifulsoup4==4.8.2
cached-property==1.5.1
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4

Click==7.0
colorama==0.4.3
future==0.18.2
gast==0.3.3
gdown==3.12.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
Jinja2==2.10.1
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0

